refilter combo after tree store additon/deletion
The combo store is a chained store linked to the main sparkpoints
store.  Values that are already selected in the dependencies tree
store need to be filtered out of the combo store because they have
already been added to the dependencies tree store.  This filter was
being done on focus, but we can't be sure a focus will actually
happen after a records has been added to or deleted from the
dependencies tree.  The combo is now filtered on afterrender (for
first filter) then after additions to and deletions from the
dependencies tree store.