Revert parentId reconfiguration
Content areas will not be hierarchical, there is no parent_id field. This component is a tree for navigation purposes -- we'll be loading a flat list of content areas from the database and decorating them with navigational subnodes dynamically.

parentId is one of the tree node "meta" fields. You can't rename them, but you could use the `mapping` field config to specify the value should go somewhere else when reading/writing from a datasource (e.g. `{ name: 'parentId', mapping: 'parent_id' }`). Specifying a field here called parentId allows us to reconfigure the meta-field, but if we don't define a field called parentId the framework will add its own.

It was configured with persist: false to indicate the field should never be written to the data source. We will later use a validation rule on parentId to enforce that the navigation child nodes we generate later can't be saved.

It's also worth noting that specifying a type for a field is only necessary if you need to force type coercion.