correct ordering of replicated fields
Clones were being added at the index position of the original field.  Re: Wrike issue 58685112, I updated this so that they would be added after the last (previous) clone.  Really we might be able to just use ownerCt.add(clone) without the index, but this seems designed to function with other fields in the same container, as long as the other fields do not also use the field replicator plugin.