Reset checked attribute on non-leaf nodes
The top level nodes are occasionally picking up a checked:true attribute
after learn creation/modification.  This resets the checked attribute to
null for nodes without a standardCode on restoreState.  This should fix
Wrike ID  58685071, but as noted in code comment, a better solution
would be to understand why the checked attroibute is being set and
prevent it from happening.